task 4 completed

+++ module01/task4/src/main/java/calculator/Calculator.java
@@ -9,7 +9,7 @@ public class Calculator {
         Printer.printToConsole("Enter an expression: ");
         Expression expression = ConsoleReader.readExpression();
 
-        int result = expression.calculate();
+        double result = expression.calculate();
         Printer.printToConsole(String.valueOf(result));
     }
 }
